+++ NEWS.md
@@ -1,10 +1,27 @@
+# metrica 1.3.0
+
+* June 13th, 2022. <br/>
+
+Adding new plotting function:
+  - density_plot, to create a plot that paints density areas of predicted and observed coordinates.
+  
+Adding new arguments to plotting functions scatter_plot, tiles_plot, and density_plot:
+  - print_eq (to print equation or not)
+  - position_metrics (to customize position of metrics_table when print_metrics = TRUE)
+  - position_eq (to customize position of equation when print_eq = TRUE)
+  - colors (for tile_plot and density_plot, to paint gradient from "low" to "high" density of points)
+
+Adding two new functions to the family of regression performance metrics:
+  - dcorr (Distance correlation) as wrapper of dcor() from energy package (https://CRAN.R-project.org/package=energy), and
+  - MIC (Maximal Information Coefficient) as wrapper of mine_stat() from minerva package (https://CRAN.R-project.org/package=minerva).
+
+# Previous versions
+
 # metrica 1.2.4
 
 * May 27th, 2022. <br/>
 
 Development update fixing minor a bug on the NSE() function to estimate the Nash-Sutcliffe model efficiency that produce only values of NSE=1.00 on v1.23.
-
-# Previous versions
 
 # metrica 1.2.3